Extract more road stuff
