+++ lib/index.ts
@@ -121,7 +121,7 @@ export * from './src/createAlchemy'
 export * from './src/createMagic'
 export * from './src/createPubRumour'
 export * from './src/createRing'
-export * from './src/createRoad'
+export * from './src/createRoadEncounter'
 export * from './src/createRoleplayQuestion'
 export * from './src/createTrap'
 export * from './src/defineRollDataGetter'
@@ -160,11 +160,15 @@ export * from './src/weightedRandomFetcher'
 export * from './src/weightRandom'
 export * from './src/worldType'
 
+export * from './roads/assignRoad'
+export * from './roads/createRoad'
+export * from './roads/createRoadName'
+export * from './roads/findExistingRoad'
+export * from './roads/roadMaterialTypes'
+export * from './roads/roadNameProperNouns'
 export * from './roads/roads'
 export * from './roads/RoadType'
 export * from './roads/roadTypes'
-export * from './roads/assignRoad'
-export * from './roads/findExistingRoad'
 
 export * from './tavern/_common'
 export * from './tavern/createTavernName'
